Part 4 final refactor

+++ part3/blog_backend/controllers/blogs.js
@@ -62,9 +62,17 @@ blogsRouter.put('/:id', async (request, response) => {
 blogsRouter.delete(`/:id`, async (request, response, next) => {
   const blog = await Blog.findById(request.params.id)
 
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
   const user = await User.findById(decodedToken.id)
+
+  if (!user) {
+    return response.status(401).json({ error: 'user not found' })
+  }
 
   if ( blog.user.toString() === user.id.toString() ) {
     await Blog.findByIdAndDelete(request.params.id)
